Fixed 115378 and 115501. Also make a kill command smart enough not to do the kill if PID is 0.

+++ cnd.debugger.gdb/src/org/netbeans/modules/cnd/debugger/gdb/proxy/GdbLogger.java
@@ -49,7 +49,9 @@ public class GdbLogger {
         File tmpfile;
         try {
             tmpfile = File.createTempFile("gdb-cmds", ".log"); // NOI18N
-            tmpfile.deleteOnExit();
+	    if (!Boolean.getBoolean("gdb.console.savelog")) { // NOI18N - This lets me save logs
+		tmpfile.deleteOnExit();
+	    }
             logFile = new FileWriter(tmpfile);
         } catch (IOException ex) {
             logFile = null;
